Implemented the very simple UnionFind algorithm
Implemented the version taking just a number as parameter for the
constructor.
This version don't use real nodes. It expect that each nodes had a
identifier indexed from 0.

I plan to identify nodes with an index to solve a problem that will
occur due to the fact that we don't wont to use a list with not a O(1)
access. So I have to modify the in and out value of Edge.